feat: implement demo tenant schema isolation with dynamic connections and tenant-aware models
- add IS_DEMO flag and tenancy provider bootstrap
- introduce TenantConnectionService for schema creation, migrations, seeding, and connection lifecycle
- replace demo sqlite session setup with per-visitor postgres schema provisioning
- add tenant-aware Lucid adapter using AsyncLocalStorage connection context
- enable demo middleware only when demo mode is active
- update seeders to use model relations for pivot records and berth crane assignments
- add pivot timestamps support and updated_at columns for assigned role tables
- fix vessel primary key to use imoNumber and update vessel numeric fields to floats

+++ database/migrations/1775586978782_create_vessels_table.ts
@@ -11,10 +11,10 @@ export default class extends BaseSchema {
       table.string('flag_state').notNullable()
       table.string('type').notNullable()
 
-      table.integer('gross_tonnage').notNullable()
-      table.integer('loa').notNullable()
-      table.integer('beam').notNullable()
-      table.integer('max_draft').notNullable()
+      table.float('gross_tonnage').notNullable()
+      table.float('loa').notNullable()
+      table.float('beam').notNullable()
+      table.float('max_draft').notNullable()
       table.integer('teu_capacity').notNullable()
 
       table
